requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.1
Django==3.2.23
djangorestframework==3.12.4
geographiclib==2.0
geopy==2.4.0
idna==3.10
numpy==1.24.0
pandas==2.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2024.2
requests==2.31.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.1
urllib3==2.3.0
